requirementsCuda.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
idna==3.11
importlib_resources==6.5.2
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
lightning==2.5.6
lightning-utilities==0.15.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mpmath==1.3.0
multidict==6.7.0
narwhals==2.10.2
networkx==3.5
nibabel==5.3.2
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==11.3.0
plotly==6.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
pure_eval==0.2.3
pydicom==3.0.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.6
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
simpleitk==2.5.2
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
sympy==1.13.1
tcia_utils==3.2.1
threadpoolctl==3.6.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchmetrics==1.8.2
torchvision==0.21.0+cu124
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
wcwidth==0.2.14
yarl==1.22.0
